Robot player with mouse look
