+++ Symmetrical Tribble/Assets/LocomotionSMB.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 1414c4d78946571ba9b4fe329cac9f38
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
