+++ Symmetrical Tribble/Assets/Metal_box_HQ.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 189b941c11ce02bd383e57f55ce90ac0
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
