+++ Symmetrical Tribble/Assets/Metal_box_HQ/Meshes.meta
@@ -1,0 +1,9 @@
+fileFormatVersion: 2
+guid: d0fa9a178d2b12741b05d6a9ceff8662
+folderAsset: yes
+timeCreated: 1435242847
+licenseType: Store
+DefaultImporter:
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
