+++ Symmetrical Tribble/Assets/Metal_box_HQ/Meshes/Materials.meta
@@ -1,0 +1,9 @@
+fileFormatVersion: 2
+guid: 5d3ca587dc162624080770fb132f4dc4
+folderAsset: yes
+timeCreated: 1435242852
+licenseType: Store
+DefaultImporter:
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
