+++ Symmetrical Tribble/Assets/Metal_box_HQ/Meshes/Materials/Box_1.mat.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 492ba10cce4e6af40b15f28111c6f4f9
+timeCreated: 1435242853
+licenseType: Store
+NativeFormatImporter:
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
