+++ Symmetrical Tribble/Assets/Metal_box_HQ/Meshes/Materials/Box_2.mat.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: e52d95e85f21a764eb63af347311a6bb
+timeCreated: 1435242852
+licenseType: Store
+NativeFormatImporter:
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
