+++ Symmetrical Tribble/Assets/Metal_box_HQ/Textures.meta
@@ -1,0 +1,9 @@
+fileFormatVersion: 2
+guid: d6143ee5d5b43e24c91b990cc347a9ef
+folderAsset: yes
+timeCreated: 1435242847
+licenseType: Store
+DefaultImporter:
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
