+++ Symmetrical Tribble/Assets/_Completed-Game/Animations/Run_0.fbx.meta
@@ -225,17 +225,17 @@ ModelImporter:
       firstFrame: 8
       lastFrame: 62
       wrapMode: 0
-      orientationOffsetY: 0.71
+      orientationOffsetY: 1.16
       level: 0
       cycleOffset: 0
       loop: 0
       hasAdditiveReferencePose: 0
       loopTime: 1
       loopBlend: 1
-      loopBlendOrientation: 0
-      loopBlendPositionY: 0
+      loopBlendOrientation: 1
+      loopBlendPositionY: 1
       loopBlendPositionXZ: 0
-      keepOriginalOrientation: 0
+      keepOriginalOrientation: 1
       keepOriginalPositionY: 1
       keepOriginalPositionXZ: 0
       heightFromFeet: 0
