--- Symmetrical Tribble/Assets/_Completed-Game/Prefabs/Pick Up.prefab.meta
@@ -1,8 +1,0 @@
-fileFormatVersion: 2
-guid: 83a9f6978de8f4ed58b9e0d00a781d5d
-timeCreated: 1480430409
-licenseType: Store
-NativeFormatImporter:
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
