--- Symmetrical Tribble/Assets/_Completed-Game/Scripts/LocomotionSB.cs
@@ -1,20 +1,0 @@
-using UnityEngine;
-
-public class LocomotionSMB : StateMachineBehaviour
-{
-
-    public float m_Damping = 0.15f;
-
-    private readonly int m_HashHorizontalPara = Animator.StringToHash("Horizontal");
-    private readonly int m_HashVerticalPara = Animator.StringToHash("Vertical");
-
-    override public void OnStateUpdate(Animator animator, AnimatorStateInfo stateInfo, int layerIndex) {
-        float horizontal = Input.GetAxis("Horizontal");
-        float vertical = Input.GetAxis("Vertical");
-
-        Vector2 input = new Vector2(horizontal, vertical).normalized;
-
-        animator.SetFloat(m_HashHorizontalPara, input.x, m_Damping, Time.DeltaTime);
-        animator.SetFloat(m_HashVerticalPara, input.y, m_Damping, Time.DeltaTime);
-    }
-}
