--- Symmetrical Tribble/Assets/_Completed-Game/Scripts/LocomotionSB.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: 5f9b04e58941d3163aa8f2a55a7291ef
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
